예약 시스템 기능 개선 및 구조 변경
