+++ templates/parking_lots.html
@@ -3,17 +3,19 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>주차장 목록 울랄라~</title>
+    <title>주차장 목록</title>
     <link rel="stylesheet" href="{{ url_for('static', filename='bootstrap/bootstrap.min.css') }}">
 </head>
 <body>
     <div class="container">
-        <h2 class="text-center mt-3">주차장 목록 울랄라~</h2>
+        <h2 class="text-center mt-3">주차장 목록</h2>
+
+        {% if parking_lots.items %}
         <ul class="list-group mt-3">
             {% for lot in parking_lots.items %}
             <li class="list-group-item">
-                <strong>이름:</strong> 
-                <a href="{{ url_for('parkinglot.parking_lot_detail', parkinglot_id=lot.parkinglot_id) }}">
+                <strong>이름:</strong>
+                <a href="{{ url_for('parkinglot_bp.parking_lot_detail_route', parkinglot_id=lot.parkinglot_id) }}">
                     {{ lot.parkinglot_name }}
                 </a>
                 <br>
@@ -26,22 +28,29 @@
         <nav aria-label="Page navigation" class="mt-3">
             <ul class="pagination">
                 {% if parking_lots.has_prev %}
-                    <li class="page-item"><a class="page-link" href="{{ url_for('parkinglot.parking_lots', page=parking_lots.prev_num) }}">이전</a></li>
+                    <li class="page-item">
+                        <a class="page-link" href="{{ url_for('parkinglot_bp.parking_lots_route', page=parking_lots.prev_num) }}">이전</a>
+                    </li>
                 {% endif %}
                 {% for num in parking_lots.iter_pages() %}
                     {% if num %}
                         <li class="page-item {% if num == parking_lots.page %}active{% endif %}">
-                            <a class="page-link" href="{{ url_for('parkinglot.parking_lots', page=num) }}">{{ num }}</a>
+                            <a class="page-link" href="{{ url_for('parkinglot_bp.parking_lots_route', page=num) }}">{{ num }}</a>
                         </li>
                     {% else %}
                         <li class="page-item disabled"><span class="page-link">…</span></li>
                     {% endif %}
                 {% endfor %}
                 {% if parking_lots.has_next %}
-                    <li class="page-item"><a class="page-link" href="{{ url_for('parkinglot.parking_lots', page=parking_lots.next_num) }}">다음</a></li>
+                    <li class="page-item">
+                        <a class="page-link" href="{{ url_for('parkinglot_bp.parking_lots_route', page=parking_lots.next_num) }}">다음</a>
+                    </li>
                 {% endif %}
             </ul>
         </nav>
+        {% else %}
+        <p class="text-center text-muted mt-4">❌ 등록된 주차장이 없습니다.</p>
+        {% endif %}
     </div>
 </body>
 </html>
